move AttributeProviderService.h to the controllers directory
